+++ farmer/ncc/validations/__init__.py
@@ -1,4 +1,3 @@
-from .evaluate import evaluate, save_show_results
 from .confidence_plot import confidence_plot
 from .confusion_visualize import confusion_visualize
 from .pca import pca
